+++ x/marker/simulation/proposals_test.go
@@ -4,8 +4,6 @@ import (
 	"math/rand"
 	"testing"
 
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-
 	"github.com/stretchr/testify/require"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 
@@ -29,7 +27,20 @@ func TestProposalContents(t *testing.T) {
 	accounts := simtypes.RandomAccounts(r, 3)
 
 	// execute ProposalContents function
-	weightedProposalContent := simulation.ProposalContents(keeper.NewKeeper(app.AppCodec(), app.GetKey(types.ModuleName), app.GetSubspace(types.ModuleName), app.AccountKeeper, app.BankKeeper, app.AuthzKeeper, app.FeeGrantKeeper, app.TransferKeeper, app.GetKey(banktypes.StoreKey)))
+	weightedProposalContent := simulation.ProposalContents(
+		keeper.NewKeeper(
+			app.AppCodec(),
+			app.GetKey(types.ModuleName),
+			app.GetSubspace(types.ModuleName),
+			app.AccountKeeper,
+			app.BankKeeper,
+			app.AuthzKeeper,
+			app.FeeGrantKeeper,
+			app.TransferKeeper,
+			app.AttributeKeeper,
+			app.NameKeeper,
+		),
+	)
 	require.Len(t, weightedProposalContent, 7)
 
 	w0 := weightedProposalContent[0]